This program is meant to be a game based off of the Phoenix Wright Series. It is a roleplaying game in which the player plays as a lawyer trying to get his client out of trouble. It uses three functions aside from the main function to help get the job done. The testimony function is basically where the player inputs where he/she thinks the contradictions lie. The courtRecord function shows all of the evidence accrued thus far. The gameOver function runs the game over protocall, basically you lost and can retry, or just close the program. The player should be able to see all lines of dialogue, and the end result will be the player getting a NOT GUILTY verdict.
First, sorry for any typos, its a long program. Second, while running the program, I noticed a lot of spacing issues even though there was nothing wrong with the code. Additionally, sometimes the getline function wouldn't opperate properly, and the program would proceed to the next line of dialogue. So if it looks off for you, it's because it would not opperate properly for me.